+++ src/screens/SplashScreen.tsx
@@ -2,12 +2,11 @@ import { useNavigation, NavigationProp, StackActions } from '@react-navigation/n
 import React, { useEffect } from 'react';
 import { View, Text } from 'react-native';
 import { useSelector } from 'react-redux';
-import { RootState } from './store';
 
 const SplashScreen: React.FC = () => {
   const navigation = useNavigation<NavigationProp<any>>();
-  const token = useSelector((state: RootState) => state.auth.token);
-  const expiredAt = useSelector((state: RootState) => state.auth.expiredAt);
+  const token = useSelector((state: { auth: { token: string } }) => state.auth.token);
+  const expiredAt = useSelector((state: { auth: { expiredAt: string } }) => state.auth.expiredAt);
 
   useEffect(() => {
     const timer = setTimeout(() => {
